update 0.0.10
Tag system added: the currently and only tag is the overlay tag but more will come as needed for adding tags to game features
the new block system now has built in blocks and every block uses a string instead of a number. the blocks are not to be passed around that much to not create much performance reduction but instead passed around by a string.
There is now a JSON file for block data with passed in tags
rewrote the rendering system for support for the new tag system ie:
the tag overlay will now render a passed in overlay name